+++ CHANGELOG.rst
@@ -28,8 +28,7 @@ see TODO.md for list of TODO items
 
   * added IsFileValidator
 
-  * added support to use prompt_toolkit for get_input. By default is used. Added option to use it or not. Also can
-    have a bottom toolbar shown. Finally, can send a key registry to set key stroke actions.
+  * changed parameter names on Regex cleaner to match re.sub parameter names
 
 * v0.2.12: